220808 add validate

+++ ssg-jojo/src/main/java/com/cloneproject/ssgjojo/categoryLv1/service/CategoryLv1ServiceImple.java
@@ -3,11 +3,13 @@ package com.cloneproject.ssgjojo.categoryLv1.service;
 import com.cloneproject.ssgjojo.categoryLv1.domain.CategoryLv1;
 import com.cloneproject.ssgjojo.categoryLv1.repository.ICategoryLv1Repository;
 import com.cloneproject.ssgjojo.categoryLv1.service.ICategoryLv1Service;
+import com.cloneproject.ssgjojo.categoryLv2.domain.CategoryLv2;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Slf4j
 @RequiredArgsConstructor
@@ -28,7 +30,18 @@ public class CategoryLv1ServiceImple implements ICategoryLv1Service {
 
     @Override
     public CategoryLv1 editCategory(CategoryLv1 categoryLv1) {
-        return iCategoryLv1Repository.save(categoryLv1);
+
+        Optional<CategoryLv1> temp = iCategoryLv1Repository.findById(categoryLv1.getId());
+        if(temp.isPresent()) {
+            return iCategoryLv1Repository.save(CategoryLv1.builder()
+                    .id(temp.get().getId())
+                    .lv1name(categoryLv1.getLv1name())
+                    .lv1imgpath(categoryLv1.getLv1imgpath())
+                    .build());
+        }
+
+        return null;
+
     }
 
     @Override
@@ -38,6 +51,10 @@ public class CategoryLv1ServiceImple implements ICategoryLv1Service {
 
     @Override
     public void deleteCategory(Long id) {
-        iCategoryLv1Repository.deleteById(id);
+
+        Optional<CategoryLv1> temp = iCategoryLv1Repository.findById(id);
+        if(temp.isPresent()) {
+            iCategoryLv1Repository.deleteById(id);
+        }
     }
 }
