+++ ssg-jojo/src/main/java/com/cloneproject/ssgjojo/categoryLv4/service/CategoryLv4ServiceImple.java
@@ -1,5 +1,8 @@
 package com.cloneproject.ssgjojo.categoryLv4.service;
 
+import com.cloneproject.ssgjojo.categoryLv2.domain.CategoryLv2;
+import com.cloneproject.ssgjojo.categoryLv2.dto.CategoryLv2Dto;
+import com.cloneproject.ssgjojo.categoryLv3.domain.CategoryLv3;
 import com.cloneproject.ssgjojo.categoryLv4.domain.CategoryLv4;
 import com.cloneproject.ssgjojo.categoryLv4.dto.CategoryLv4Dto;
 import com.cloneproject.ssgjojo.categoryLv3.repository.ICategoryLv3Repository;
@@ -8,7 +11,9 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
+import javax.swing.text.html.Option;
 import java.util.List;
+import java.util.Optional;
 
 @Slf4j
 @RequiredArgsConstructor
@@ -21,11 +26,15 @@ public class CategoryLv4ServiceImple implements ICategoryLv4Service {
     @Override
     public CategoryLv4 addCategory(CategoryLv4Dto categoryLv4Dto) {
 
-//      return iCategoryRepositoryS.save(categoryS);
-        return iCategoryLv4Repository.save(CategoryLv4.builder()
-                .lv4name(categoryLv4Dto.getLv4name())
-                .categoryLv3(iCategoryLv3Repository.findById(categoryLv4Dto.getCategoryLv3()).get())
-                .build());
+        Optional<CategoryLv3> temp = iCategoryLv3Repository.findById(categoryLv4Dto.getCategoryLv3());
+        if(temp.isPresent()) {
+            return iCategoryLv4Repository.save(CategoryLv4.builder()
+                    .lv4name(categoryLv4Dto.getLv4name())
+                    .categoryLv3(iCategoryLv3Repository.findById(categoryLv4Dto.getCategoryLv3()).get())
+                    .build());
+        }
+
+        return null;
     }
 
     @Override
@@ -36,12 +45,18 @@ public class CategoryLv4ServiceImple implements ICategoryLv4Service {
     @Override
     public CategoryLv4 editCategory(CategoryLv4Dto categoryLv4Dto) {
 
-//      return iCategoryRepositoryS.save(categoryS);
-        return iCategoryLv4Repository.save(CategoryLv4.builder()
-                .id(categoryLv4Dto.getId())
-                .lv4name(categoryLv4Dto.getLv4name())
-                .categoryLv3(iCategoryLv3Repository.findById(categoryLv4Dto.getCategoryLv3()).get())
-                .build());
+        Optional<CategoryLv4> categoryLv4 = iCategoryLv4Repository.findById(categoryLv4Dto.getId());
+        Optional<CategoryLv3> categoryLv3 = iCategoryLv3Repository.findById(categoryLv4Dto.getCategoryLv3());
+
+        if(categoryLv3.isPresent() && categoryLv4.isPresent()) {
+            return iCategoryLv4Repository.save(CategoryLv4.builder()
+                    .id(categoryLv4Dto.getId())
+                    .lv4name(categoryLv4Dto.getLv4name())
+                    .categoryLv3(iCategoryLv3Repository.findById(categoryLv4Dto.getCategoryLv3()).get())
+                    .build());
+        }
+
+        return null;
     }
 
     @Override
@@ -51,6 +66,9 @@ public class CategoryLv4ServiceImple implements ICategoryLv4Service {
 
     @Override
     public void deleteCategory(Long id) {
-        iCategoryLv4Repository.deleteById(id);
+        Optional<CategoryLv4> temp = iCategoryLv4Repository.findById(id);
+        if(temp.isPresent()) {
+            iCategoryLv4Repository.deleteById(id);
+        }
     }
 }
